+++ src/SCexport/ExportLogDialog.Designer.cs
@@ -127,7 +127,7 @@ namespace SCaddins.SCexport
             this.tabPage2.Location = new System.Drawing.Point(4, 22);
             this.tabPage2.Name = "tabPage2";
             this.tabPage2.Padding = new System.Windows.Forms.Padding(3);
-            this.tabPage2.Size = new System.Drawing.Size(511, 380);
+            this.tabPage2.Size = new System.Drawing.Size(511, 357);
             this.tabPage2.TabIndex = 1;
             this.tabPage2.Text = "Warnings";
             this.tabPage2.UseVisualStyleBackColor = true;
@@ -140,7 +140,7 @@ namespace SCaddins.SCexport
             this.warnings.Dock = System.Windows.Forms.DockStyle.Fill;
             this.warnings.Location = new System.Drawing.Point(3, 3);
             this.warnings.Name = "warnings";
-            this.warnings.Size = new System.Drawing.Size(505, 374);
+            this.warnings.Size = new System.Drawing.Size(505, 351);
             this.warnings.TabIndex = 0;
             this.warnings.UseCompatibleStateImageBehavior = false;
             this.warnings.View = System.Windows.Forms.View.Details;
@@ -161,7 +161,7 @@ namespace SCaddins.SCexport
             this.tabPage3.Location = new System.Drawing.Point(4, 22);
             this.tabPage3.Name = "tabPage3";
             this.tabPage3.Padding = new System.Windows.Forms.Padding(3);
-            this.tabPage3.Size = new System.Drawing.Size(511, 380);
+            this.tabPage3.Size = new System.Drawing.Size(511, 357);
             this.tabPage3.TabIndex = 2;
             this.tabPage3.Text = "Errors";
             this.tabPage3.UseVisualStyleBackColor = true;
@@ -174,7 +174,7 @@ namespace SCaddins.SCexport
             this.errors.Dock = System.Windows.Forms.DockStyle.Fill;
             this.errors.Location = new System.Drawing.Point(3, 3);
             this.errors.Name = "errors";
-            this.errors.Size = new System.Drawing.Size(505, 374);
+            this.errors.Size = new System.Drawing.Size(505, 351);
             this.errors.TabIndex = 0;
             this.errors.UseCompatibleStateImageBehavior = false;
             this.errors.View = System.Windows.Forms.View.Details;
@@ -206,6 +206,7 @@ namespace SCaddins.SCexport
             this.Controls.Add(this.tabControl1);
             this.Icon = ((System.Drawing.Icon)(resources.GetObject("$this.Icon")));
             this.Name = "ExportLogDialog";
+            this.StartPosition = System.Windows.Forms.FormStartPosition.CenterParent;
             this.Text = "Export Summary";
             this.TopMost = true;
             this.tabControl1.ResumeLayout(false);